requirements.txt
pypdf
langchain_community
langchain_chroma
langchain_openai
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib